Refactoring, se agregó README.md. App ya en funcionamiento
