+++ requirements.txt
@@ -1,0 +1,5 @@
+chromadb==0.4.18
+keybert==0.8.3
+langchain==0.0.350
+llama_index==0.9.14.post3
+Requests==2.31.0
